tasks/requirements.txt
celery==5.3.4
pydantic==1.10.9
python-dotenv==1.0.0
redis==5.0.1
